ui/requirements.txt
streamlit==1.48.1
websockets==15.0.1
websocket-client==1.8.0
loguru==0.7.3